pyproject.toml
[tool.poetry]
name = "aml-sdk-demo"
version = "0.1.0"
description = ""
authors = ["Brittany Pugh <brittanypugh@microsoft.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
azure-ai-ml = "^1.30.0"
azure-identity = "^1.14.0"
numpy = "^1.24.0"
pandas = "^2.1.0"
scikit-learn = "^1.3.0"
matplotlib = "^3.7.0"
seaborn = "^0.12.0"
mlflow = "^2.8.0"
azure-storage-blob = "^12.19.0"
python-dotenv = "^1.0.0"
pyyaml = "^6.0"
jupyterlab = "^4.0.0"
plotly = "^5.17.0"
pydantic = "^2.12.3"
torch = { version = "2.7.0+cpu", source = "pytorch-cpu"}
azureml-mlflow = "^1.60.0.post1"
joblib = "^1.5.2"

[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
mypy = "^1.6.0"
pre-commit = "^3.5.0"
ruff = "^0.1.0"
pyright = "^1.1.0"
rich = "^13.0.0"

[tool.ruff]
exclude = ["typings"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "TID"]

[tool.ruff.format]
docstring-code-format = true

[tool.pyright]
typeCheckingMode = "off"
venvPath = "."
venv = ".venv"
exclude = ["typings", ".venv", "notebooks", ]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
